Make the bridge depend on generated code
