+++ flutter_ffi_plugin/example/native/hub/src/bridge/interface.rs
@@ -2,28 +2,13 @@
 
 use crate::tokio;
 use rinf::externs::lazy_static::lazy_static;
+use rinf::SimpleCell;
 use std::cell::RefCell;
-use std::sync::Arc;
-use std::sync::Mutex;
 
 #[cfg(not(target_family = "wasm"))]
 pub use super::interface_os::*;
 #[cfg(target_family = "wasm")]
 pub use super::interface_web::*;
-
-/// This contains a message from Dart.
-/// Optionally, a custom binary called `blob` can also be included.
-/// This type is generic, and the message
-/// can be of any type declared in Protobuf.
-pub struct DartSignal<T> {
-    pub message: T,
-    pub blob: Option<Vec<u8>>,
-}
-
-/// This is a mutable cell type that cannot be shared across threads.
-pub type SimpleCell<T> = RefCell<Option<T>>;
-/// This is a mutable cell type that can be shared across threads.
-pub type SharedCell<T> = Arc<Mutex<SimpleCell<T>>>;
 
 #[cfg(not(target_family = "wasm"))]
 lazy_static! {
