+++ flutter_ffi_plugin/lib/src/exports.dart
@@ -1,2 +1,1 @@
 export 'interface_os.dart' if (dart.library.html) 'interface_web.dart';
-export 'interface.dart';
