+++ rust_crate/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod externs;
+
+mod common;
+pub use common::*;
